% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.gfi.R
\name{assoc.gfi}
\alias{assoc.gfi}
\title{Generalized affiliation index}
\usage{
assoc.gfi(M1, M2, fr = T, sym = F, erase.diag = T)
}
\arguments{
\item{M1}{a square adjacency matrix representing individual interactions or associations.}

\item{M2}{a square adjacency matrix representing individual values of confounding factors.}

\item{fr}{if \emph{true}, it considers the argument M1 as an adjacency matrix representing interaction frequencies between individuals. 
Otherwise, it considers the argument M1 as an adjacency matrix representing associations between individuals.}

\item{sym}{if \emph{true}, it considers the argument M1 as an adjacency matrix representing symmetric interactions/associations.}

\item{erase.diag}{if \emph{true}, it omits the diagonal of the matrix.}
}
\value{
a square adjacency matrix representing the generalized affiliation index between individuals.
}
\description{
Computes generalized affiliation indices based on a matrix of interactions or associations and a confounding factor.
}
\details{
Generalized affiliation indices allow to control for individual associations by a given confounding factor (such as temporal or spatial overlaps, gregariousness, social unit membership, kinship...).
The principle is to perform a Generalized Linear Regression (GLR) on both matrices (one representing the individual interactions/associations and the other one representing the confounding factor)
and to use GLR residuals as association indices. For an adjacency matrix representing individual interactions, the GLR belongs to the Poisson family. For an adjacency matrix representing individual associations, the GLR belongs to the Binomial family. 
High positive values suggest strong associations between two individuals and negative values suggest avoidance between two individuals.
}
\references{
Whitehead, H., & James, R. (2015). Generalized affiliation indices extract affiliations from social network data. Methods in Ecology and Evolution, 6(7), 836-844.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
