% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assoc_mat_full}
\alias{assoc_mat_full}
\title{Association indexes}
\usage{
assoc_mat_full(Mgbi, method)
}
\arguments{
\item{method}{a string indicating the type of association matrix:
\itemize{
\item 'sri' for Simple ratio index: \eqn{x/x+yAB+yA+yB}
\item 'hw' for Half-weight index: \eqn{x/x+yAB+1/2(yA+yB)}
\item 'sr' for Square root index:\eqn{x/sqr((x+yAB+yA)(x+yAB+yB))}
}}

\item{Mfbi}{a group by individual matrix.}
}
\value{
A square matrix of association according to association index choose:
}
\description{
Individual associations indexes.
}
\details{
Association indexes allow to handle  heterogeneity of time of observation.
}
\references{
Whitehead, H. A. L. (1997). Analysing animal social structure. Animal behaviour, 53(5), 1053-1067.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
