% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.df.R
\name{check.df}
\alias{check.df}
\title{Data frame ANT check}
\usage{
check.df(df)
}
\value{
A vector if argument df is a data frame and/or a list of data frames
}
\description{
Check that the argument is a data frame or a list of data frames
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
