% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.id.R
\name{check.id}
\alias{check.id}
\title{Data frame ANT ID check}
\usage{
check.id(df)
}
\description{
Check that the ID column of a data frame or a list is set as factorof data frames
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
