% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.matrix.R
\name{check.mat}
\alias{check.mat}
\title{Matrix ANT check}
\usage{
check.mat(M)
}
\arguments{
\item{M}{a square adjacency matrix or a list of square adjacency matrices}
}
\value{
Nothing if M is a square adjacency matrix or a list of square adjacencies matrices.
}
\description{
check if the input is a square adjacency matrix or a list of squares adjacency matrices.
}
\details{
tests if M is not a data frame and if M is a matrix and if it is a square matrix. If the input is a list then it check for all elements of the list if they are not a data frame and if they are a square adjacency matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
