% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.socprog.R
\name{convert.socprog}
\alias{convert.socprog}
\title{Converts Socprog data frame}
\usage{
convert.socprog(df, id, scan, sep = ";")
}
\arguments{
\item{df}{a 'dyadic' or 'group' data frame socprog format}

\item{id}{an integer or string indicating the column of ids}

\item{scan}{an integer or string vector indicating the column of the scans}

\item{sep}{a character indicating the type of sepration between individuals inside one scan}
}
\value{
a data frame in socprog format.
}
\description{
Converts socprog 'Group mode' data frame (one scan per line with several individuals observed during the scan)
}
\examples{
head(sim.socprog)
convert.socprog(sim.socprog,id=3,scan=c(1,2),sep=";")
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
