% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT.mcmc.troph_function.R
\name{CT.mcmc.troph}
\alias{CT.mcmc.troph}
\title{Continuous Time MCMC Estimation function for Trophallaxis data}
\usage{
CT.mcmc.troph(a, b, c, d, r, q, tau, theta, data, init, Time, n.mcmc,
  save = F)
}
\value{
(1) - estimates of X, gamma, lambda, P
}
\description{
The purpose of this function is to find MCMC generated estimates of 
(1) -
}
\examples{

}
