% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.col.findId.R
\name{df.col.findId}
\alias{df.col.findId}
\title{Finds a data frame index}
\usage{
df.col.findId(df, label_name)
}
\arguments{
\item{df}{a data frame in which to find the index of a specific column(s).}

\item{label_name}{a character or numeric vector indicating the column name or index respectively.}
}
\value{
an numeric vector corresponding to the column index that matches argument label_name.
}
\description{
Finds the data frame index of a column from the name of the column or his index.
}
\details{
This function allows the user to select on or several columns according to their name(s) or their index(es).
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
