% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.create.single.R
\name{df.create.single}
\alias{df.create.single}
\title{Creates an empty data frame.}
\usage{
df.create.single(M, names = TRUE)
}
\arguments{
\item{M}{a matrix.}

\item{names}{if \emph{true} then a column in the data frame is added with the names of the matrix columns.}
}
\value{
A data frame of N rows based on the number of rows in argument M.
}
\description{
Creates an empty data frame with as many rows or columns as in the corresponding matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
