% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.ctrlFactor.R
\name{df.ctrlFactor}
\alias{df.ctrlFactor}
\title{Creates a column that merges multiple columns}
\usage{
df.ctrlFactor(df, control)
}
\arguments{
\item{df}{a data frame.}

\item{control}{a numeric or character vector representing one or more columns that are combined into one column of control factors.}
}
\value{
The input data frame with an additional column named 'control', representing the control factors.
}
\description{
Creates a new column in the input data frame corresponding to the combination of given control factors.
}
\details{
Control factors are used in permutation approaches to constrain the permutation between those factors.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
