% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.to.gbi.focal.R
\name{df.to.gbi.focal}
\alias{df.to.gbi.focal}
\title{Converts a data frame of individual associations into a group by individual matrix.}
\usage{
df.to.gbi.focal(df, focal, alters, ctrl)
}
\arguments{
\item{df}{a data frame in which to include a control column.}

\item{scan}{a numeric or character vector representing one or more columns used as scan factors.}

\item{id}{a numeric or character vector indicating the column holding ids of individuals.}
}
\value{
The same data frame input with an extra column named 'control', representing the control factors.
}
\description{
Converts a data frame of individual associations into a group by individual matrix.
}
\details{
Control factors are used in permutation approaches to constrain their permutations.
}
\examples{
df.to.gbi.focal(df=sim.focal.undirected,focal='focal', alters='alter',ctrl='nfocal')
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
