% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.to.gbi.R
\name{df.to.gbi}
\alias{df.to.gbi}
\title{Data frame to GBI.}
\usage{
df.to.gbi(df, scan, id)
}
\arguments{
\item{df}{a data frame of individual associations.}

\item{scan}{a numeric or character vector representing one or more columns used as scan factors.}

\item{id}{a numeric or character vector indicating the column holding ids of individuals.}
}
\value{
A group by individual matrix.
}
\description{
Converts a data frame of individual associations into a group by individual matrix.
}
\details{
Several association indices are coputed on GBI.
}
\examples{
head(sim.grp)
df.to.gbi(sim.grp,scan=c('location','time'),id='ID')
}
\seealso{
\code{\link{assoc.indices}}
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
