% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{df_to_gbi}
\alias{df_to_gbi}
\title{Group by individual matrix.}
\usage{
df_to_gbi(d, col_scan, col_id, uIds, uScans)
}
\arguments{
\item{d}{a data frame.}

\item{col_scan}{an integer indicating the column number holding the information of the scans.}

\item{col_id}{an integer indicating the column number holding the individuals observations.}

\item{uIds}{a character vector holding all the ids of the individuals.}

\item{uScans}{a character vector holding all the ids of the scans.}
}
\value{
A group by individual matrix.
}
\description{
Convert a data frame into a group by individual matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales
}
\keyword{internal}
