% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.mcmc.troph_function.R
\name{DT.mcmc.troph}
\alias{DT.mcmc.troph}
\title{Discrete Time MCMC Estimation function for Trophallaxis data}
\usage{
DT.mcmc.troph(y.data, ant.file, title, a, b, c, d, theta, states = 2, n.mcmc,
  delta.t, hours, param.start)
}
\arguments{
\item{y.data, }{ant.file, title, a, b, theta, states, n.mcmc, delta.t}
}
\value{
(1) - estimates of X, lambda, P
         (2) - 2x2 visual of estimates over time (runs)
}
\description{
The purpose of this function is to find MCMC generated estimates of 
(1) - the state (X_t = high/low troph rates) of the colony at time t
(2) - the specific rates of interaction (lambda) of each state 
(3) - the probability of moving from one state to another (P matrix)
}
\examples{
theta = matrix(data = c(90, 10, 10, 90), nrow = 2, ncol = 2, byrow = T) 
out.high = DT.mcmc.troph(data = high.y, title = 'High Density', a = 5, b = 2, 
theta = theta, states = 2, n.mcmc = 3000)
}
