% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.pen.cov.mcmc.troph_function.R
\name{DT.pen.cov.mcmc.troph}
\alias{DT.pen.cov.mcmc.troph}
\title{Penalized Discrete Time MCMC Estimation function for Trophallaxis data}
\usage{
DT.pen.cov.mcmc.troph(penalty, covariate, y.data, states, ant.file, hours, a, b,
  c, d, tau, n.mcmc, seconds, fig.save, start)
}
\arguments{
\item{y.data, }{ant.file, title, a, b, c, d, theta, states, n.mcmc, delta.t}
}
\value{
(1) - estimates of betas, alpha, X, lambda, gamma, P
         (2) - 2x2 visual of estimates over time (runs)
         (3) - color block state switching graph
}
\description{
The purpose of this function is to find MCMC generated estimates of 
(1) - the state (X_t = high/low troph rates) of the colony at time t
(2) - the specific rates of interaction (lambda) of each state 
(3) - the specific rates of state switching (gamma) for the ants
(3a) - the probability of moving from one state to another (P matrix) 
       generated from gamma values.
}
\examples{



}
