% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.pen.mcmc.troph.expprior_function.R
\name{DT.pen.mcmc.troph.expprior}
\alias{DT.pen.mcmc.troph.expprior}
\title{Penalized Discrete Time MCMC Estimation function for Trophallaxis data}
\usage{
DT.pen.mcmc.troph.expprior(penalty, y.data, states, ant.file, hours, a, b, c, d,
  tau, tau.pen, n.mcmc, seconds, fig.save, start)
}
\arguments{
\item{y.data, }{ant.file, title, ?, theta, states, n.mcmc, delta.t}
}
\value{
(1) - estimates of X, lambda, gamma, P
         (2) - 2x2 visual of estimates over time (runs)
         (3) - color block state switching graph
}
\description{
This function differes from DT.pen.mcmc.troph only in that it features 
an Exponential prior on the gamma values instead of Half Normal priors.
}
\details{
The purpose of this function is to find MCMC generated estimates of 
(1) - the state (X_t = high/low troph rates) of the colony at time t
(2) - the specific rates of interaction (lambda) of each state 
(3) - the specific rates of state switching (gamma) for the ants
(3a) - the probability of moving from one state to another (P matrix) 
       generated from gamma values.
}
\examples{
theta = matrix(data = c(90, 10, 10, 90), nrow = 2, ncol = 2, byrow = T) 
out.high = DT.pen.mcmc.troph()


}
