% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_model_functions.R
\name{DT_pencov_mcmc}
\alias{DT_pencov_mcmc}
\title{Penalized Discrete Time MCMC Estimation function for 
Trophallaxis data with Covariates}
\usage{
DT_pencov_mcmc(penalty, covariate, starts_data, states, ant_file, chamber,
  hours, title, a, b, c, d, tau, n_mcmc, delta_t, fig_save, fig_path, fig_name,
  start)
}
\arguments{
\item{penalty}{Penalty parameter for the switching rates 
(alpha/beta in model write- up).}

\item{covariate}{Vector of second by second covariate data 
used to estimate parameters.}

\item{starts_data}{Data file of number of trophallais 
interactions starting per delta_t time increments.}

\item{states}{Number of unobserved states, generally 2 ("low",
"high").}

\item{ant_file}{Data file of ant trophallaxis interaction 
information, including Location and start/end times.}

\item{hours}{Hours of ant obseravtions reflected in data 
files.}

\item{title}{Title of diagnostic plots used to check model 
results.}

\item{a}{Hyperprior for rates of starting interactions during 
"low" state. (lambda_L)}

\item{b}{Hyperprior for rates of starting interactions during 
"low" state. (lambda_L)}

\item{c}{Hyperprior for rates of starting interactions during 
"high" state. (lambda_H)}

\item{d}{Hyperprior for rates of starting interactions during 
"high" state. (lambda_H)}

\item{tau}{Tuning parameter for proposal function.}

\item{n_mcmc}{Number of MCMC iterations.}

\item{delta_t}{Time segments that the starting times are 
binned into.}

\item{fig_save}{If "TRUE", plots will saved in path proved as 
.jpeg files}

\item{fig_path}{Path needed to sent plot figures to folder.}

\item{fig_name}{Base name of plot files to be saved.}

\item{start}{Starting values for the chains of parameters to 
be estimated.}

\item{title}{Title of diagnostic plots.}
}
\value{
(1) - estimates of betas, alpha, X, lambda, gamma, P 
  (2) - 2x2 visual of estimates over time (runs) (3) - color 
  block state switching graph
}
\description{
The purpose of this function is to find MCMC generated 
estimates of (1) - the state (X_t = high/low troph rates) of 
the colony at time t (2) - the specific rates of interaction 
(lambda) of each state (3) - the specific rates of state 
switching (gamma) for the ants (3a) - the probability of 
moving from one state to another (P matrix) generated from 
gamma values.
}
