% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgl.to.grp.R
\name{edgl.to.grp}
\alias{edgl.to.grp}
\title{Dyadic data frame (edge list) to linear group by individual data frame}
\usage{
edgl.to.grp(df, id1, id2, scan)
}
\arguments{
\item{df}{a symetric edgelist of individual associations.}

\item{id1}{a numeric or character vector indicating the column of the first individual association.}

\item{id2}{a numeric or character vector indicating the column of the second individual association.}

\item{scan}{a numeric or character vector representing one or more columns used as scan factors.}
}
\value{
a data frame with individual associations according to scan variable(s) declared by the user.
}
\description{
Converts a dyadic data frame (edge list) to a linear group by individual data frame
}
\keyword{internal}
