% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edgl_to_matrix}
\alias{edgl_to_matrix}
\title{Edge list to matrix}
\usage{
edgl_to_matrix(df, sym)
}
\arguments{
\item{df}{a data frame contening at leas three columns with the names: 'from','to' and 'weight', corresponding respectively to the individual emiting the beaheviour, the individual receving the beahviour, and the number of time this interaction occure.}

\item{sym}{a boolean if TRUE then the interactions are concider to be symetric.}
}
\value{
A square adjacency matrix.
}
\description{
Convert an edge list to a matrix.
}
\details{
Adjacency matrix is one of the numerous representation of a social network.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
