% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbi.to.df.R
\name{gbi.to.df}
\alias{gbi.to.df}
\title{Group by individual matrix to linear data frame of association}
\usage{
gbi.to.df(gbi)
}
\arguments{
\item{gbi}{a a group (N) by individual (I) matrix}
}
\value{
a linear data frame of association.
}
\description{
Convert Linear gbi to a inear data frame of association
}
\keyword{internal}
