% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.lpEnergyDegree.R
\name{laplacian_energy_degrees}
\alias{laplacian_energy_degrees}
\title{Laplacian energy.}
\usage{
laplacian_energy_degrees(M)
}
\arguments{
\item{M}{a square adjacency matrix.}
}
\description{
Calculate the Laplacian energy.
}
\details{
Laplacian energy is the sum of squares of the eigenvalues in the Laplacian matrix.
This version use the met.degree to calculated the Laplacian energy.
}
\references{
REF laplacian !!!!!!!!!!!!!!!!

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
