% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_lapply}
\alias{list_lapply}
\title{Lapply function}
\usage{
list_lapply(input, f)
}
\arguments{
\item{input}{a list.}

\item{f}{a function to apply along the list.}
}
\value{
A list with the output type of the function.
}
\description{
C++ version of lapply function.
}
\author{
This function have been copy from Rcpp gallery: XXXX
}
\keyword{internal}
