% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_to_df}
\alias{list_to_df}
\title{List to data frame.}
\usage{
list_to_df(a)
}
\arguments{
\item{lst}{a list.}
}
\value{
A data frame with the same number of columns as the list length.
}
\description{
Convert a list of vectors into a data frame
}
\author{
This function have been copy from Rcpp Gallery: http://gallery.rcpp.org/articles/faster-data-frame-creation/
}
\keyword{internal}
