% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listDf_merge_single_column}
\alias{listDf_merge_single_column}
\title{Merge columns of a List of data frames.}
\usage{
listDf_merge_single_column(df, ldf, col)
}
\arguments{
\item{df}{a data frame to know the total size of the concatenated vector.}

\item{ldf}{a list of data frames in wich extarct a single column.}

\item{col}{an integer indicating wich column of the data frames extract. Is a R indexation.}
}
\value{
A character vector.
}
\description{
perm_dataStraem_focal subfunction to concatenate the columns of alters of a list of data frames.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
