% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.lp.R
\name{mat.lp}
\alias{mat.lp}
\title{Laplacian matrix}
\usage{
mat.lp(M)
}
\arguments{
\item{M}{a square adjacency matrix.}
}
\value{
a numeric vector.
}
\description{
Create the laplacian matrix of a network.
}
\details{
Laplacian matrix is the difference between the degrees matrix and the adjacency matrix.
}
\references{
REF laplacian !!!!!!!!!!!!!!!!

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
