% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.symetrize.R
\name{mat.symetrize}
\alias{mat.symetrize}
\title{Symmetrizes matrix}
\usage{
mat.symetrize(M)
}
\arguments{
\item{M}{a matrix}
}
\description{
Symmetrizes matrix by triangles sum.
}
\author{
Sebastian Sosa <s.sosa@live.fr>
}
\keyword{internal}
