% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.to.edgl.R
\name{mat.to.edgl}
\alias{mat.to.edgl}
\title{Matrix to edge list}
\usage{
mat.to.edgl(M, sym = F, erase.diag = T)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{sym}{if \emph{true}, it extracts the lower triangle of the matrix only.}

\item{erase.diag}{if \emph{true}, it omits diagonals.}
}
\description{
Converts a square adjacency matrix into a data frame of three columns representing an edge list. Columns are: actor, receiver and weight.
}
\examples{
mat.to.edgl(sim.m, sym = FALSE, erase.diag = TRUE)
}
\references{
Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
