% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.vectorization.R
\name{mat.vectorization}
\alias{mat.vectorization}
\title{Matrix vectorization}
\usage{
mat.vectorization(M, sym = FALSE, erase.diag = TRUE)
}
\arguments{
\item{M}{an adjacency matrix or list of adjacency matrices.}

\item{sym}{if \emph{true} will vectorize only the lower triangle.}

\item{erase.diag}{if \emph{true} will not keep the diagonal of the matrix.}
}
\value{
a numeric vector
}
\description{
Transform a matrix into a vector.
}
\details{
Transform an adjacency matrix (symetric or not) into a vector.
}
\keyword{internal}
