% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_binaryzation}
\alias{mat_binaryzation}
\title{Matrix binarization.}
\usage{
mat_binaryzation(M)
}
\arguments{
\item{M}{a data frame to know the total size of the concatenated vector.}
}
\value{
A matrix.
}
\description{
Convert matrix element higher to one in zeros.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
