% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_cols_sums}
\alias{mat_cols_sums}
\title{From Column Sums.}
\usage{
mat_cols_sums(m)
}
\arguments{
\item{m}{a matrix.}
}
\value{
An integer vector. Each elements correspond to the sum of the corresponding column.
}
\description{
Column Sums of an adjacency matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
