% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_dim}
\alias{mat_dim}
\title{Dimensions of a matrix}
\usage{
mat_dim(m)
}
\arguments{
\item{m}{a matrix.}
}
\value{
A numeric vector with two elements. The first one correspond to the number of row and the second one correspond to the number of columns.
}
\description{
Retrieve or set the dimension of Matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
