% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_filter}
\alias{mat_filter}
\title{Matrix filtering}
\usage{
mat_filter(m, threshold, replace)
}
\arguments{
\item{m}{a matrix.}

\item{threshold}{a value above wich the cells of the matrix will be keep.}

\item{replace}{a doubl eindicating the value by which replace the replace elements.}
}
\value{
A matix.
}
\description{
Filtering matrix cells according to a threshold and replace them by a bouble.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
