% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_find0}
\alias{mat_find0}
\title{Find zeros in a Matrix.}
\usage{
mat_find0(m)
}
\arguments{
\item{m}{a matrix.}
}
\value{
A data frame with two columns the first one representing the row id and the second one trepresenting the column id of the zero cell.
}
\description{
Find empty cells in a matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
