% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_isSquare}
\alias{mat_isSquare}
\title{Is square.}
\usage{
mat_isSquare(m)
}
\arguments{
\item{m}{a matrix.}
}
\value{
a boolean indicating if the matrix is square or not.
}
\description{
Test if a matrix is square or not.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
