% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_row_extract}
\alias{mat_row_extract}
\title{Extract matrix row.}
\usage{
mat_row_extract(m, x)
}
\arguments{
\item{m}{a matrix.}

\item{int}{na integer indicating the row number to extract.}
}
\value{
An integer vector corresponding the row extract.
}
\description{
Extract a specific row of a matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
