% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_rows_sums}
\alias{mat_rows_sums}
\title{From row Sums.}
\usage{
mat_rows_sums(x)
}
\arguments{
\item{m}{a matrix.}
}
\value{
An integer vector. Each elements correspond to the sum of the corresponding row.
}
\description{
row Sums of an adjacency matrix.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
