% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.troph.cov_function.R
\name{mcmc.troph.cov}
\alias{mcmc.troph.cov}
\title{MCMC Estimation function for Trophallaxis data with Covariate(s)}
\usage{
mcmc.troph.cov(y.data, ant.file, inout.file, title, a = 5, b = 2, theta,
  states = 2, n.mcmc, cov, mu.cov, sig.cov, tau, delta.t, hours)
}
\arguments{
\item{y.data, }{ant.file, inout.file, title, a, b, theta, states, n.mcmc, cov, mu.cov, sig.cov, tau, delta.t}
}
\value{
(1) - estimates of X, lambda, P
         (2) - 2x2 visual of estimates over time (runs)
}
\description{
The purpose of this function is to find MCMC generated estimates of 
(1) - the state (X_t = high/low troph rates) of the colony at time t
(2) - the specific rates of interaction (lambda) of each state 
(3) - the probability of moving from one state to another (P matrix)
This model will take in observed interaction data and covariates. Currently
time since last forager ant entered the chamber is our covariate.
}
\examples{
theta = matrix(data = c(90, 10, 10, 90), nrow = 2, ncol = 2, byrow = T) 
mu.all = c(2, -1, -0.000004)
sig.all = matrix(data = c(0.2, 0, 0, 
                         0, 0.2, 0, 
                         0, 0, 0.0002), nrow = 3, ncol = 3, byrow = T)
                         
in.out =  read.csv("Data/Colony_1_in&out_high_density_4hrs.csv")
in.out = in.out[which(in.out$Action == "enter"),]
in.out = in.out[order(in.out$time), ]

#Create vector of covariates - time since last ant entered
cov = rep(300, in.out$time[1])
for(i in 2:nrow(in.out)){
   cov = c(cov, 0:(in.out$time[i] - in.out$time[i - 1] ))
 }
 
cov = c(cov, 0:(14400 - in.out$time[nrow(in.out)]))
out.high = mcmc.troph.cov(data = high.y, title = "High Density", a = 5, b = 2, 
theta = theta, states = 2, n.mcmc = 3000, cov = cov, mu.cov = mu.all, sig.cov = sig.all)
}

