% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.troph_function.R
\name{mcmc.troph}
\alias{mcmc.troph}
\title{MCMC Estimation function for Trophallaxis data}
\usage{
mcmc.troph(data, title, a, b, theta, states, n.mcmc)
}
\arguments{
\item{data, }{title, a, b, theta, states, n.mcmc}
}
\value{
(1) - estimates of X, lambda, P
         (2) - 2x2 visual of estimates over time (runs)
}
\description{
The purpose of this function is to find MCMC generated estimates of 
(1) - the state (X_t = high/low troph rates) of the colony at time t
(2) - the specific rates of interaction (lambda) of each state 
(3) - the probability of moving from one state to another (P matrix)
}
\examples{
theta = matrix(data = c(90, 10, 10, 90), nrow = 2, ncol = 2, byrow = T) 
out.high = mcmc.troph(data = high.y, title = "High Density", a = 5, b = 2, 
theta = theta, states = 2, n.mcmc = 3000)
}

