% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.affinity.single.R
\name{met.affinity.single}
\alias{met.affinity.single}
\title{Affinity}
\usage{
met.affinity.single(M, df = NULL, dfid = NULL, binary = FALSE)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{df}{a data frame of same length of the input matrix.}

\item{binary}{a boolean, if \emph{true}, it calculates the binary version of the affinity.}
}
\value{
Integer vector of each vertices met.affinity.
}
\description{
Calculate for all the vertices the node metric call Affinity
}
\details{
Affinity is a second order metric that try to evaluate to which egos, node \emph{in} is connected. The binary version  is simply the average degree of alters of node i. The weighted version is ratio between the met.reach metric and the met.strength of node i.A high met.affinity reveals that nodes tend to be connected to alters of high degrees or strengths. Thus, both metrics provide information on node assortativity by vertex met.degree, i.e. connections between nodes with similar degrees or strengths.
}
\references{
Whitehead, H. A. L. (1997). Analysing animal social structure. Animal behaviour, 53(5), 1053-1067.

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
