% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.perm.ds.grp.R
\name{met.all}
\alias{met.all}
\title{All node metrics}
\usage{
met.all(M, df, vec)
}
\description{
Computes all network metrics in ANT package according to the character vector input
}
\keyword{internal}
