% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.alterDegree.R
\name{met.alterDegree}
\alias{met.alterDegree}
\title{Alters sum or average of met.degree or met.strength.}
\usage{
met.alterDegree(M, i, av = FALSE, binary = FALSE)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{i}{The ego in which we want to calculate his aleer met.degree.}

\item{av}{a boolean, if TRUE compute the sum of alters before making alter mean.}
}
\value{
Vector fo Alters sum or average of met.degree or met.strength values for each verteces.
}
\description{
Computer the degre of a specific node.
}
\details{
calculates average or sum of alters' degrees. By default it calculates the sum of the degrees of alters.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
