% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.assortativityCat.R
\name{met.assortativityCat}
\alias{met.assortativityCat}
\title{Assortativity}
\usage{
met.assortativityCat(M, attr, df = NULL)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{attr}{a factor or character vector indicating individuals categorical attributes appartenance.}
}
\value{
an integer representing the categorical assortativity index of the network.
}
\description{
Calculates the binary or weighted version of vertices Newman's assortativity for categorical or continuous attributes.
}
\details{
Assortativity (like the E-I index or the Moran 'I' statistic)  allows the study of homophily (preferential interaction between nodes with similar attributes) and heterophily (the preferential interaction between nodes with different attributes). Attributes can be individual characteristics such as sex or age, or individual node metrics such as the met.degree, in which case it is referred to as assortativity by vertex met.degree.
}
\references{
Newman, M. E. (2003). Mixing patterns in networks. Physical Review E, 67(2), 026126.

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
