% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.betweenness.single.R
\name{met.betweenness.single}
\alias{met.betweenness.single}
\title{Betweenness centrality}
\usage{
met.betweenness.single(m, binary = T, shortest.weight = F,
  normalization = T, sym = T, out = T, df = NULL, dfid = NULL)
}
\description{
Compute node betweenness
}
\keyword{internal}
