% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.cc.R
\name{met.cc}
\alias{met.cc}
\title{Binary global clustering coefficient}
\usage{
met.cc(M)
}
\arguments{
\item{M}{a square adjacency matrix.}
}
\value{
An integer representing the binary global clustering coefficient index of the network.
}
\description{
Calculates the binary global clustering coefficient.
}
\details{
The binary global clustering coefficient is the ratio of closed triplets to all triplets (open and closed) in a network. Triplets are three nodes connected by two or three undirected links (open or closed triplets, respectively).
}
\references{
Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
