% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.ci.single.R
\name{met.ci.single}
\alias{met.ci.single}
\title{Centralisation index}
\usage{
met.ci.single(M)
}
\arguments{
\item{M}{a square adjacency matrix or a list of squares adjacencymatrices}
}
\value{
numeric vector
}
\description{
Compute network Centralisation index
}
\details{
Centralisation index of a network is based on eigenvector centrality.
}
\references{
Pasquaretta, C., Leve, M., Claidiere, N., Van De Waal, E., Whiten, A., MacIntosh, A. J., ... & Crofoot, M. C. (2014). Social networks in primates: smart and tolerant species have more efficient networks. Scientific reports, 4, 7600.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
