% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.coutTriangles.R
\name{met.coutTriangles}
\alias{met.coutTriangles}
\title{Calculates the number of triangles in a network}
\usage{
met.coutTriangles(M)
}
\arguments{
\item{M}{an adjacency matrix.}
}
\value{
An integer representing the binary global clustering coefficient index of the network.
}
\description{
Calculates the number of triangles in a network.
}
\details{
It uses an algorithm similar to the Node-iterator-core algorithm.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
