% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.diversity.R
\name{met.diversity}
\alias{met.diversity}
\title{Nodes diversity}
\usage{
met.diversity(m, sym = TRUE)
}
\arguments{
\item{m}{a square matrix}

\item{sym}{bollean indicating if matrix must be symetrize (Nodes diversity doesn't concider directionality)}
}
\value{
A numeric vector indicating diversity for all nodes
}
\description{
Compute nodes diversity
}
\details{
Diversity of node is equal to the scaled Shannon entropy of a node incoming edges.
}
\references{
Nathan Eagle, Michael Macy and Rob Claxton: Network Diversity and Economic Development, Science 328, 1029–1031, 2010.
}
