% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.indegree.single.R
\name{met.indegree.single}
\alias{met.indegree.single}
\title{Indegree}
\usage{
met.indegree.single(M, df = NULL, dfid = NULL)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{df}{a data frame of same length of the input matrix.}

\item{dfid}{an integer indicating the column of individual ids in argument \emph{df}}
}
\value{
Integer vector of each vertices met.indegree.
}
\description{
Calculate for all the vertices the node metric call met.indegree.
}
\details{
met.indegree of vertice \emph{i} is the number of all edges receive by node \emph{i}.
}
\references{
Barrat, A., Barthelemy, M., Pastor-Satorras, R., & Vespignani, A. (2004). The architecture of complex weighted networks. Proceedings of the National Academy of Sciences of the United States of America, 101(11), 3747-3752.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
