% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.lp.single.R
\name{met.lp.single}
\alias{met.lp.single}
\title{Symetric Laplacian centrality}
\usage{
met.lp.single(M, df = NULL, dfid = NULL, binary = FALSE)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{df}{a data frame of same length of the input matrix.}

\item{dfid}{an integer indicating the column of individual ids in argument \emph{df}}

\item{binary}{a boolean indicating if the binary version of the laplacian centrality has to be computed.}
}
\value{
a numeric vector.
}
\description{
Calculate the symetric version of the Laplacian centrality for each verteces.
}
\details{
Laplacian centrality is the drop in the Laplacian energy of the graph when the vertex is removed.
This version uses the degrees (for the binary version ) or the strength (for the weigthed version) to calculate laplacian centrality.
}
\references{
REF laplacian !!!!!!!!!!!!!!!!

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
