% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.lpcB.R
\name{met.lpcB}
\alias{met.lpcB}
\title{Binary symetric Laplacian centrality}
\usage{
met.lpcB(M)
}
\arguments{
\item{M}{a square adjacency matrix.}
}
\value{
a numeric vector.
}
\description{
Calculate the binary symetric Laplacian centrality for each verteces.
}
\details{
Laplacian centrality is the drop in the Laplacian energy of the graph when the vertex is removed
This version uses the degrees to calculate laplacian centrality
}
\references{
REF laplacian !!!!!!!!!!!!!!!!

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
