% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.lpcentEvcent.R
\name{met.lpcentEvcent}
\alias{met.lpcentEvcent}
\title{Laplacian centrality}
\usage{
met.lpcentEvcent(M, binary = FALSE)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{binary}{if if \emph{true}  then calculate the directed binary version of Laplacian centrality}
}
\value{
a numeric vector.
}
\description{
Calculate the directed weighted or binary Laplacian centrality for each verteces.
}
\details{
Laplacian centrality is the drop in the Laplacian energy of the graph when the vertex is removed
This version uses the eigenvalues to calculate laplacian centrality
}
\references{
REF laplacian !!!!!!!!!!!!!!!!

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
