% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.lpcW.R
\name{met.lpcW}
\alias{met.lpcW}
\title{Wheigthed symetric Laplacian centrality}
\usage{
met.lpcW(M)
}
\arguments{
\item{M}{a square adjacency matrix.}
}
\description{
Calculate the wheigthed symetric Laplacian centrality for each verteces.
}
\details{
Laplacian centrality is the drop in the Laplacian energy of the graph when the vertex is removed
This version uses the degrees to calculate laplacian centrality
}
\references{
REF laplacian !!!!!!!!!!!!!!!!

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
