% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.lpEnergyEigen.R
\name{met.lpEnergyEigen}
\alias{met.lpEnergyEigen}
\title{Laplacian energy.}
\usage{
met.lpEnergyEigen(M)
}
\arguments{
\item{M}{a square adjacency matrix.}
}
\description{
Calculate the Laplacian energy.
}
\details{
Laplacian energy is the sum of squares of the eigenvalues in the Laplacian matrix.
This version use the eigenvalues to calculated the Laplacian energy.
}
\references{
REF laplacian !!!!!!!!!!!!!!!!

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
