% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{met}
\alias{met}
\title{ANT Graphical User Interface (GUI)}
\usage{
met(M, df = NULL, dfid = NULL)
}
\arguments{
\item{M}{a square adjacency matrix or a list of square adjacency matrices or output from ANT functions of type perm'.}

\item{df}{a data frame of same length as in the input matrix or a list of data frames of same length as the correspondoing matrix in argument M.}

\item{dfid}{an integer or a string indicating the column of individual ids in argument \emph{df}}
}
\description{
Graphical user interface for the calculation of node metrics.
}
\details{
returns the data frame with node network metrics, if weighted =TRUE:
\itemize{
\item Weighted met.degree (WD)
\item Weighted met.indegree (WID)
\item Weighted met.outdegree (WOD)
\item Weighted met.met.ri (met.met.ri)
\item Weighted met.evcent (WEi)
\item Weighted clustering coefficient (WCC)
\item Weighted closeness (WCL)
\item Weighted betweenness (WDB)
\item Weighted betweenness (WUB)
\item Weighted met.disparity (met.disparity)
}
if weighted =NULL, it returns binary network metrics:
\itemize{
\item Binary met.degree (BD)
\item Binary met.indegree (BID)
\item Binary met.outdegree (BOD)
\item Binary met.met.ri (met.met.ri)
\item Binary met.evcent (BEi)
\item Binary clustering coefficient (BCC)
\item Binary closeness (BCL)
\item Binary betweenness (DBB)
\item Binary betweenness (UBB)
\item Binary met.evcent community (e.i. modularity) (BEiM)
}
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
