% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.reach.single.R
\name{met.reach.single}
\alias{met.reach.single}
\title{Reach centrality}
\usage{
met.reach.single(M, df = NULL, dfid = NULL, return.strength = FALSE)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{df}{a data frame of same length of the input matrix.}

\item{dfid}{an integer indicating the column of individual ids in argument \emph{df}.}

\item{return.strength}{a boolean that indicate if strength have to be return. Possiblility only available if if argument \emph{df} and \emph{dfid} are both \emph{NULL} Default value is \emph{FALSE}.}
}
\value{
Integer vector of each vertices R-Index values.
}
\description{
Calculate for all the vertices the node metric call reach.
}
\details{
reach is the sum of the product of all the ego and alter's strengths and the alters' degrees
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
