% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.ri.single.R
\name{met.ri.single}
\alias{met.ri.single}
\title{R-Index}
\usage{
met.ri.single(M, df = NULL, dfid = NULL)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{df}{a data frame of same length of the input matrix.}

\item{dfid}{an integer indicating the column of individual ids in argument \emph{df}}
}
\value{
Integer vector of each vertices R-Index values. If the adjacency matrix is binary it will return the binari R-index. If the adjacency matrix is weigthed it will return the binari R-index.
}
\description{
Calculate for all the vertices the node metric call R-Index.
}
\details{
R-Index of vertice \emph{i} is the outstrength of node \emph{i} divided by the met.strength  of node \emph{i}. Such node metric atempt to measur if an individual have more tendency to receive or emit edges.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
