% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.robustness.R
\name{met.robustness}
\alias{met.robustness}
\title{Metric robustness}
\usage{
met.robustness(df, subsampling = c(5, 10, 20, 30, 40, 50),
  metric = "met.strength", assoc.indices = FALSE, actor, receiver,
  scan, id, index = "sri", ...)
}
\arguments{
\item{df}{a data frame of individual associations}

\item{subsampling}{a vector of integers indicating the percentage of data to remove}

\item{metric}{the network metric to compute}

\item{assoc.indices}{a bollean indicating if association indices have to be use}

\item{actor}{If argument assoc.indices is FALSE fill this argument, an integer or a string indicating the column of the individuals performing the behaviour.}

\item{receiver}{If argument assoc.indices is FALSE fill this argument, an integer or a string indicating the column of the individuals receiving the behaviour.}

\item{scan}{If argument assoc.indices is TRUE fill this argument, ana numeric or character vector representing one or more columns used as scan factors.}

\item{id}{If argument assoc.indices is TRUE fill this argument, ana numeric or character vector indicating the column holding ids of individuals.}
}
\value{
3 elements:
\itemize{
\item A matrix in which each columns represent a node metric variation throough bootstraping, with the first row representing the original metric.
\item A summary of bootstrap distribution for each nodes.
\item A plot of the 
}
}
\description{
Perform a matrix boostrapping approach to estimate the confidence intervals surrounding each pairwise association.
}
\details{
This process evaluate networks metrics robustness by performing a boostrapp process to remove a certain percent of the data collected and recomputing the network metric of intereste. This have been usde in Balasubramaniam et al. 2018 to asses sensitivty of sampling effor of global network metrics in primates networks.
}
\examples{
df = sim.focal.directed
test = met.robustness(df, actor = 'actor', receiver = 'receiver', metric = 'met.strength')
# objects return by the function
test$metrics
test$summary
test$plot
# Examples with metric extra arguments
met.robustness(df, actor = 'actor', receiver = 'receiver', metric = 'met.affinity')
met.robustness(df, actor = 'actor', receiver = 'receiver', metric = 'met.affinity', sym = FALSE)
# Examples with association data
test2 = met.robustness(df = sim.grp, assoc.indices = TRUE, scan = c('time', 'location'), id = 'ID', metric  = 'met.strength')
# Example of how to test global network metric robustness by removing 10\% of the observations simulated 100 times
test = met.robustness(df, subsampling = rep(10,100), actor = 'actor', receiver = 'receiver', metric = 'met.diameter')
}
\references{
Balasubramaniam, K. N., Beisner, B. A., Berman, C. M., De Marco, A., Duboscq, J., Koirala, S., ... & Ogawa, H. (2018). The influence of phylogeny, social style, and sociodemographic factors on macaque social network structure. American journal of primatology, 80(1), e22727.
}
\author{
Sebastian Sosa
}
