% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{met_ei}
\alias{met_ei}
\title{Eigenvector centrality}
\usage{
met_ei(m)
}
\arguments{
\item{m}{a matrix.}
}
\value{
A numeric vector corresponding to the Eigenvector centrality of each nodes.
}
\description{
compute the eigenvecto centrality for each vertices through Armadillo c++ library.
}
\details{
EigenvectorCpp centrality is the first non-negative met.eigenvector value obtained through the linear transformation of an adjacency matrix. This centrality measure quantifies not only a node connectedness, but also the connections of the nodes to whom it is connected. Thus, a node can have a high met.evcent value by having a high met.degree or met.strength, or by being connected to nodes that have high degrees or strengths.
}
\references{
Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
