% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{met_strength}
\alias{met_strength}
\title{strength}
\usage{
met_strength(M)
}
\arguments{
\item{M}{a square adjacency matrix.}
}
\value{
Integer vector of each vertices outstrentgh.

Integer vector of each vertices outstrentgh.
}
\description{
Calculate for all the vertices the node metric call strentgh.

Calculate for all the vertices the node metric call outstrentgh.
}
\details{
strength of vertice \emph{i} is all edges //' @title Outstrength
}
\references{
Barrat, A., Barthelemy, M., Pastor-Satorras, R., & Vespignani, A. (2004). The architecture of complex weighted networks. Proceedings of the National Academy of Sciences of the United States of America, 101(11), 3747-3752.

Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzales.
}
\keyword{internal}
