% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metric_global_shortestDetails}
\alias{metric_global_shortestDetails}
\title{Geodesic distances details}
\usage{
metric_global_shortestDetails(disMap)
}
\description{
Compute the geodesic distances of a matrix and retunr detials of the paths
}
\keyword{internal}
