% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metric_global_shortestDetailsBasedBetween}
\alias{metric_global_shortestDetailsBasedBetween}
\title{Geodesic distances details for betweenness calculation}
\usage{
metric_global_shortestDetailsBasedBetween(disMap)
}
\description{
Compute the geodesic distances of a matrix and retunr detials of the paths for betweenness calculation
}
\keyword{internal}
