% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metric_global_shortestPath}
\alias{metric_global_shortestPath}
\title{Geodesic distances}
\usage{
metric_global_shortestPath(disMap)
}
\description{
Compute the geodesic distances of a matrix
}
\keyword{internal}
