% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.vis.R
\name{net.vis}
\alias{net.vis}
\title{Network vizualisation}
\usage{
net.vis(
  m,
  df,
  id,
  shape = NULL,
  size = NULL,
  color = NULL,
  label = NULL,
  n.col1 = "blue",
  n.col2 = "red",
  n.size = 1,
  e.width = 1,
  background = NULL,
  solver = NULL,
  n.shape = NULL,
  viewer = FALSE
)
}
\arguments{
\item{m}{A square adjacency matrix.}

\item{df}{A data frame of node attributes.}

\item{id}{A string or integer indicating the column in the data frame
with nodes ids.}

\item{shape}{A string or integer indicating the column in the data frame
according to the shape nodes must have.}

\item{size}{A string or integer indicating the column in the data frame
according to the size nodes must have.}

\item{color}{A string or integer indicating the column in the data frame
according to the color nodes must have.}

\item{label}{A string or integer indicating the column in the data frame
with the labels to use.}

\item{n.col1}{A string indicating the first color value with which to start
the color gradiant.}

\item{n.col2}{A string indicating the last color value with which to end
the color gradiant.}

\item{n.size}{An integer indicating the size of nodes.}

\item{e.width}{An integer indicating the width of edges.}

\item{background}{A string of html color code.}

\item{solver}{A string indicating the spatialization solver to use among
the following options: 'barnesHut', 'repulsion', 'hierarchicalRepulsion',
'forceAtlas2Based'.}

\item{n.shape}{An optional string vector if you want to set nodes' shapes.
It must be unique values of same length as argument df in shape column.}

\item{viewer}{A boolean indicating if plots must be displayed in R Studio
viewer.}
}
\value{
The network representation in your browser.
}
\description{
This function draw network.
}
\examples{
sim.df= met.eigen(sim.m,sim.df)
head(sim.df)
net.vis(sim.m, sim.df, id = 'id', shape = 'sex', n.shape = c('circle', 'triangle'),
  size = 'eigen', color = 'age')
}
\author{
Sebastian Sosa, Jérôme Pansanel
}
