% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeskMetrics2.R
\name{nodeskMetrics2}
\alias{nodeskMetrics2}
\title{Interpret ANT GUI output}
\usage{
nodeskMetrics2(M, option, df = NULL, dfid = NULL)
}
\arguments{
\item{M}{a matrix}

\item{option}{a numeric vector issue from nodeskMetrics function.}

\item{df}{a data frame of same row number as the input matrix}
}
\description{
Interpret tcltk \emph{nodeskMetrics} function output
}
\keyword{internal}
