% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification.R
\name{notification}
\alias{notification}
\title{TCLTK notification}
\usage{
notification(text = "Analysis over")
}
\description{
TCLTK popup with text message ("Analysis over" by default)
}
\author{
Sebastian Sosa
}
