% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses_functions.R
\name{penalty_diagnosis}
\alias{penalty_diagnosis}
\title{Diagnoses of Penalized HMM Model}
\usage{
penalty_diagnosis(mcmc_matrix, Time, fig_path, fig_name, penalty)
}
\arguments{
\item{mcmc_matrix}{}

\item{penalty}{}
}
\description{
Diagnoses of Penalized HMM Model
}
