% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.dataStream.focal.R
\name{perm.dataStream.focal}
\alias{perm.dataStream.focal}
\title{Converts a data frame of individual associations into a group by individual matrix.}
\usage{
perm.dataStream.focal(
  df,
  focal,
  scan,
  alters,
  nperm,
  progress = TRUE,
  method = "sri"
)
}
\arguments{
\item{df}{a data frame in which to include a control column.}

\item{scan}{a numeric or character vector representing one or more columns used as scan factors.}

\item{id}{a numeric or character vector indicating the column holding ids of individuals.}
}
\value{
The same data frame input with an extra column named 'control', representing the control factors.
}
\description{
Converts a data frame of individual associations into a group by individual matrix.
}
\details{
Control factors are used in permutation approaches to constrain their permutations.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
