% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.edgl.R
\name{perm.edgl}
\alias{perm.edgl}
\title{Edge list link permutations on an edgelist}
\usage{
perm.edgl(edgelist, perm)
}
\arguments{
\item{edgelist}{an edgelist}

\item{perm}{number of permutations to perform.}
}
\value{
a list of edge lists with permuted links. The first element of the list is the original edge list.
}
\description{
Permutes the links of an edge list represented by a data frame.
}
\details{
edge list links permutations can be used to create random networks based on the observed network. Such permutation method is useful when analysing patterns of interactions such as assortativity.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
