% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.degree.single.R
\name{perm.met.degree.single}
\alias{perm.met.degree.single}
\title{Link perm keeping structure on a single matrix}
\usage{
perm.met.degree.single(M, perm)
}
\arguments{
\item{M}{a square adjacency matrix or a list of square adjacency matrix.}

\item{perm}{number of perm wanted}
}
\value{
\itemize{
\item A list of prmuted matrices, if M is a single square adjacency matrix.
\item A list of list if M is a list of square adjacency matrix.
}
}
\description{
Permute row and/or columns of a matrix in order to keep links structure constant during perm.
}
\details{
Permutation of link and keeping link structure allow to study?????????
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
