% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.net.degree.R
\name{perm.net.degree}
\alias{perm.net.degree}
\title{Link permutation keeping the structure}
\usage{
perm.net.degree(M, nperm)
}
\arguments{
\item{M}{a square adjacency matrix or a list of square adjacency matrices.}

\item{nperm}{number of permutations wanted.}
}
\value{
\itemize{
\item A list of permuted matrices, if M is a single square adjacency matrix.
\item A list of a list of permuted matrices if M is a list of square adjacency matrices.
}
}
\description{
Permutes the rows and/or the columns of a matrix to maintain the link structure during the permutation process.
}
\details{
Link permutation maintaining the link structure, is usually used for matrix correlation or regression.
}
\examples{
t=perm.net.degree(sim.m,nperm=10)
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
