% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.net.nl.str.R
\name{perm.net.nl.str.single}
\alias{perm.net.nl.str.single}
\title{Nodes labels permutation keeping network structure}
\usage{
perm.net.nl.str.single(df, labels)
}
\arguments{
\item{df}{a data frame whith nodes informations}

\item{labels}{a numeric or string vector indicating the columns (labels) to permute.}
}
\value{
a list of data frames of length nperm + 1 with the first element of the list beeing the original data frame and the other elements, the permuted ones.
}
\description{
Permute node metrics while keeping their dependency
}
\details{
Permute network links weigths while keep same network structure (density, modularity, binary global clustering coefficient).
}
\author{
Sebastian Sosa
}
\keyword{internal}
