% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.nodeLabel.R
\name{perm.nodeLabel}
\alias{perm.nodeLabel}
\title{Node label permutations}
\usage{
perm.nodeLabel(df, labels, nperm, progress = TRUE)
}
\arguments{
\item{df}{a dataframe or a list of data frames.}

\item{labels}{column names or numbers in which to permute the data}

\item{nperm}{Number of permutations.}

\item{progress}{a boolean indicating if the permutation process has to be visible.}
}
\value{
A list of data frames with the specified column permuted.
}
\description{
Performs node label permutations

Node label permutation is one of the methods used to build null models.
}
\references{
Whitehead, H. A. L. (1997). Analysing animal social structure. Animal behaviour, 53(5), 1053-1067.

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
