% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.redo.R
\name{perm.redo}
\alias{perm.redo}
\title{Repeats a permutation according to random factors}
\usage{
perm.redo(df, labels, ctrl)
}
\arguments{
\item{df}{a list of data frame with each data frame representing a subset of the random factors (e.g. groups and/or periods)}

\item{labels}{a numeric or character vector representing the labels to permute.}

\item{ctrl}{a numeric or character vector representing one or more columns that will be use as control facors.}
}
\value{
a data frame in which the data frames of the list are merged, each data frame being permuted according to the input argument labels.
}
\description{
Performs node label permutations for GLMM by intra random factor label permutations.
}
\details{
This is a function to repeat a permutation according to random factors. It allows to handle GLMM warnings or errors that may appear due to the permutation approach.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
