% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_matVec}
\alias{perm_matVec}
\title{Vectorize matrix permutation}
\usage{
perm_matVec(vec, n, rand)
}
\arguments{
\item{vec}{a vumeric matrix}

\item{n}{number of columns of the original matrix}

\item{rand}{a numeric vector of randomized column and rows indeces}
}
\value{
A vector
}
\description{
Permute rows and columns of a vectorized matrix
}
\details{
It's the equivalent of R m[rand, rand] function. Columns and rows indexatons are in R format.
}
\author{
Sebastian Sosa.
}
\keyword{internal}
