% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_net_weigths}
\alias{perm_net_weigths}
\title{Edgelist weigths permutations}
\usage{
perm_net_weigths(df, sym, nperm, progress)
}
\description{
Post-network permutations on links weights.
}
\keyword{internal}
