% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_nodeLabels}
\alias{perm_nodeLabels}
\title{Node label permutations.}
\usage{
perm_nodeLabels(df, label, nperm, progress)
}
\arguments{
\item{df}{a data frame.}

\item{label}{a numeric vector indicating the number of individuals.}

\item{nperm}{an integer indicating the number of permutations to perform.}

\item{progress}{a boolean indicating if you wich to see the progression of the permutations.}
}
\value{
A list of data frames, each ones holding a different permutation.
}
\description{
Post-network permutations of the type 'node labels'.
}
\details{
Node label permutations is a post-network permutations approach that swap the labels of the nodes (e.g. degree). It is use on direct interactions behaviours.
}
\references{
Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa,Ivan Puga-Gonzales.
}
\keyword{internal}
