% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_vec_factor}
\alias{perm_vec_factor}
\title{Permute factor vector.}
\usage{
perm_vec_factor(vec)
}
\arguments{
\item{x}{a numeric vector.}

\item{progress}{a boolean indicating if you wich to see the progression of the permutations.}
}
\value{
A list of vectors, each ones holding a different permutation.
}
\description{
Permutes elements of a vector.
}
\details{
Permuting a vector can serve to realize node label permutation.
}
\references{
Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa,Ivan Puga-Gonzales.
}
\keyword{internal}
