% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_visual_functions.R
\name{plot_states}
\alias{plot_states}
\title{Pst MCMC State Switching Visualization}
\usage{
plot_states(mcmc_matrix, fig_path, fig_name)
}
\arguments{
\item{mcmc_matrix}{}

\item{fig_path}{Path needed to sent plot figures to folder.}

\item{fig_name}{Base name of plot files to be saved.}
}
\description{
Pst MCMC State Switching Visualization
}
