% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.dist.R
\name{post.dist}
\alias{post.dist}
\title{Histogram of posterior distribution}
\usage{
post.dist(v_perm, ncols = NULL, nrows = NULL)
}
\description{
plot muliples histograms of posterior distribution
}
\keyword{internal}
