% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prep.inout.data_function.R
\name{prep.inout.data}
\alias{prep.inout.data}
\title{Data Preparation for Ant In/Out Data File}
\usage{
prep.inout.data(data, delta.t, hours)
}
\arguments{
\item{data,}{delta.t, hours}
}
\value{
(1) List file for binned covariate (forager arrival) data.
}
\description{
The purpose of this function is to take a .csv file with
entrance/exit data and format it for use in mcmc model.
}
\examples{
prep.inout.data(inout.high.4, 60, 4)
}

