% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.troph.data_function.R
\name{prep.troph.data}
\alias{prep.troph.data}
\title{Data Preparation for Ant Trophallaxis Data File}
\usage{
prep.troph.data(data, delta.t)
}
\arguments{
\item{data, }{delta.t}
}
\value{
(1) One or three (depending on high or low density, 
            respectively), file lists with amount of interaction
            within binned time chunk (delta.t).
}
\description{
The purpose of this function is to take in .csv files 
and apply necissary changes to the data format. Changes include 
removing redundant entries, sorting by entrance times, and bining data
into smaller chunks.
}
\examples{
 prep.troph.data(high4, 60)


}
