% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{prep_inout_data}
\alias{prep_inout_data}
\title{Data Preparation for Ant In/Out Data File}
\usage{
prep_inout_data(data, delta_t, hours)
}
\arguments{
\item{data}{}

\item{delta_t}{}

\item{hours}{}
}
\value{
(1) List file for binned covariate (forager arrival) data.
}
\description{
The purpose of this function is to take a .csv file with
entrance/exit data and format it for use in mcmc model.
}
\examples{
prep.inout.data(inout.high.4, 60, 4)
}

