% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.ds.focal.cum.R
\name{redo.ds.focal.cum}
\alias{redo.ds.focal.cum}
\title{Function for cumulative permutations for symmetric behaviour through focal sampling}
\usage{
redo.ds.focal.cum(df, focal, alters, ctrl, nperm, method)
}
\value{
list of square association index matrices.
}
\description{
Perform cumulative data stream permutations for scan sampling. Is the core function for the permutations.
}
\keyword{internal}
