% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.ds.focal.glm.R
\name{redo.ds.focal.glm}
\alias{redo.ds.focal.glm}
\title{Focal Data stream Recursive function for error found in permutations}
\usage{
redo.ds.focal.glm(
  family,
  formula,
  new.perm,
  gbi,
  gbi2,
  oda,
  odf,
  target.metrics,
  focal,
  ctrl,
  alters,
  index,
  start,
  control,
  model,
  method,
  x,
  y,
  contrasts,
  ...
)
}
\value{
a model object.
}
\description{
Performs cumulative data stream permutations for gambit of the group until the glmm found no error or warnings
}
\keyword{internal}
