% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.perm.ds.focal.R
\name{redo.ds.focal.glmm}
\alias{redo.ds.focal.glmm}
\title{Focal Data stream Recursive function for error found in permutations}
\usage{
redo.ds.focal.glmm(
  family,
  formula,
  new.perm,
  gbi,
  gbi2,
  oda,
  odf,
  target.metrics,
  focal,
  ctrl,
  alters,
  index,
  fam,
  ...
)
}
\description{
Performs cumulative data stream permutations for gambit of the group until the glmm found no error or warnings
}
\keyword{internal}
