% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.ds.focal.lm.R
\name{redo.ds.focal.lm}
\alias{redo.ds.focal.lm}
\title{Focal Data stream Recursive function for error found in permutations}
\usage{
redo.ds.focal.lm(
  formula,
  new.perm,
  gbi,
  gbi2,
  oda,
  odf,
  target.metrics,
  focal,
  ctrl,
  alters,
  index,
  model,
  method,
  x,
  y,
  qr,
  singular.ok,
  contrasts,
  ...
)
}
\description{
Performs cumulative data stream permutations for gambit of the group until the glmm found no error or warnings
}
\keyword{internal}
