% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.perm.ds.grp.R
\name{redo.perm.ds.grp.cum}
\alias{redo.perm.ds.grp.cum}
\title{Function for cumulative permutations on a data frame of associations}
\usage{
redo.perm.ds.grp.cum(df, Scan, method, control_factor, nperm)
}
\arguments{
\item{df}{a data frame of individuals associations}
}
\value{
A list of two elements with all the permutations : 1) gbi or control gbi if control_factor in null or not respectivelly; 2) Association matrix
}
\description{
Performs cumulative data stream permutations for gambit of the group.
}
\keyword{internal}
