% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.perm.ds.grp.R
\name{redo.perm.ds.grp.cum.scd}
\alias{redo.perm.ds.grp.cum.scd}
\title{cumulative permutations on GBI or GBI with control factors}
\usage{
redo.perm.ds.grp.cum.scd(GBI, method, nperm, control_factor)
}
\arguments{
\item{GBI}{gbi or control gbi if control_factor in null or not respectivelly}
}
\value{
A list of two elements with all the permutations : 1) gbi or control gbi if control_factor in null or not respectivelly; 2) Association matrix
}
\description{
Performs cumulative data stream permutations for gambit of the group.
}
\details{
not need of data frame of individuals associations, permutation are directly done inside the gbi, and it return a new gbi or control gbi if control_factor in null or not respectivelly with the permutations
}
\keyword{internal}
