% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.net.nl.str.R
\name{redo.perm.net.nl.str}
\alias{redo.perm.net.nl.str}
\title{Redo Nodes labels permutation keeping network structure}
\usage{
redo.perm.net.nl.str(df, labels, rf)
}
\description{
Permute node metrics while keeping their dependency. this function help to redo the permutations within stat. functions
}
\keyword{internal}
