% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{redo_perm_dataStream1_focal}
\alias{redo_perm_dataStream1_focal}
\title{Data Stream gambit of the group Permutations without control factor.}
\usage{
redo_perm_dataStream1_focal(M, M2, nperm, method)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{nperm}{an integer indicating the number of permutations to perform.}

\item{progress}{a boolean indicating if you wich to see the progression of the permutations.}
}
\value{
A list of Group By Individual matrices according to each scans perform.
}
\description{
Pre-network permutation on association data of gambit of the group type without control factor.
}
\details{
Data stream permutations is a pre-network permutations approach. It is use on association data based on the gambit of the group. This permutations functunction is made for  data collected of the type of 'gambit of the group' and without control factors
}
\references{
Whitehead, H. A. L. (1997). Analysing animal social structure. Animal behaviour, 53(5), 1053-1067.

Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Ivan Puga-Gonzales, Sebastian Sosa.
}
\keyword{internal}
