% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{redo_perm_dataStream_ControlFactor}
\alias{redo_perm_dataStream_ControlFactor}
\title{Data Stream gambit of the group cumulative permutations with control factor.}
\usage{
redo_perm_dataStream_ControlFactor(GBIList, M, nperm, CumSizesGbis, method)
}
\arguments{
\item{GBIList}{a list of gbis split according to the control(s) factors.}

\item{nperm}{an integer indicating the number of permutations to perform.}

\item{CumSizesGbis}{???}
}
\description{
Cumulative pre-network permutation on association data of gambit of the group type with control factor.
}
\details{
Data stream permutations is a pre-network permutations approach. It is use on association data based on the gambit of the group. This permutations functunction is made for  data collected of the type of 'gambit of the group' and without control factors
}
\references{
Whitehead, H. A. L. (1997). Analysing animal social structure. Animal behaviour, 53(5), 1053-1067.

Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Ivan Puga-Gonzales, Sebastian Sosa.
}
\keyword{internal}
