% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{redo_perm_dataStream_focal}
\alias{redo_perm_dataStream_focal}
\title{Data Stream Focal Sampling cumulative permutations without control factor.}
\usage{
redo_perm_dataStream_focal(df, ldf1, nperm, col_focal, col_alters, lengthList)
}
\arguments{
\item{df}{a data frame.}

\item{nperm}{an integer indicating the number of permutations to perform.}

\item{col_focal}{an integer indicating the column holding the id of the focal.}

\item{col_alters}{an integer indicating the column holding the id of the alters.}

\item{lengthList}{an integer vector range from 0 to length of the list (c++ indexation).}

\item{ldf2}{a list of data frames split by individuals focals.}

\item{id_obs}{a numeric vector with the number of different ids.}
}
\value{
A list of data frames, each ones holding a different permutation.
}
\description{
Pre-network cumulative permutation on association data of  Focal Sampling type without control factor.
}
\details{
Data stream permutations is a pre-network permutations approach. It is use on association data based on the gambit of the group. This permutations functunction is made for data collected of the type of 'focal sampling' and without control factors
}
\references{
Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa,Ivan Puga-Gonzales.
}
\keyword{internal}
