% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regspline.R
\name{regspline}
\alias{regspline}
\title{Multivariate Spline Regression with Automatic Knot Selection}
\usage{
regspline(
  formula,
  data,
  type = c("linear", "quadratic", "cubic", "quartic"),
  max_knots = 3
)
}
\arguments{
\item{formula}{formula}

\item{data}{data frame}

\item{type}{spline type: "linear","quadratic","cubic","quartic"}

\item{max_knots}{maximum number of knots per variable}
}
\description{
Multivariate Spline Regression with Automatic Knot Selection
}
