% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.uncertainty.R
\name{sampling.met}
\alias{sampling.met}
\title{Metric uncertainty}
\usage{
sampling.met(df, nperm, metric = "met.strength", assoc.indices = FALSE,
  actor = NULL, receiver = NULL, scan = NULL, id = NULL,
  index = "sri", ...)
}
\arguments{
\item{df}{a data frame of individual inetaractions or associations}

\item{metric}{the network metric to compute}

\item{M}{a square adjacency matrix}

\item{subsampling}{a vector of integers indicating the percentage of data to remove}
}
\value{
3 elements:
\itemize{
\item A matrix in which each columns represent a node metric variation throough bootstraping, with the first row representing the original metric.
\item A summary of bootstrap distribution for each nodes.
\item A plot of the 
}
}
\description{
Perform a matrix boostrapping approach to estimate the confidence intervals surrounding each pairwise association.
}
\details{
This process evaluate networks metrics uncertainty by performing a boostrapp process with replacement on the matrix of association and recomputing the network metric of intereste.
}
\examples{
test = sampling.met(sim.m, 100, metric = 'met.strength')
# objects return by the function
test$metrics
test$summary
test$plot
# Example with metric extra arguments
sampling.met(sim.m, 100, metric = 'met.affinity', binary = FALSE)
sampling.met(sim.m, 100, metric = 'met.affinity', binary = TRUE)
# Example with individuals asosciations
sampling.met(sim.grp, scan = c('location','time'), id='ID')
}
\references{
Lusseau, D., Whitehead, H., & Gero, S. (2009). Incorporating uncertainty into the study of animal social networks. arXiv preprint arXiv:0903.1519.
}
\author{
Sebastian Sosa
}
