% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.fill.research_function.R
\name{scale.fill.research}
\alias{scale.fill.research}
\title{Research scale fill for ggplot2}
\usage{
\method{scale}{fill.research}(theme = "ants", tech_key = list(ants =
  c("#bc5356", "#bcb953", "#538bbc", "#53bc84", "#8453bc")))
}
\description{
research scale fill
}
\examples{
scale_fill_tech(theme="ants")
}
\author{
Meriidth Bartley
}

