% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.df.R
\docType{data}
\name{sim.df}
\alias{sim.df}
\title{Simulated data frame}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
sim.df
}
\value{
a data frame.
}
\description{
This is a simulated data frame of individual characteristics
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
