% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.focal.directed.R
\docType{data}
\name{sim.focal.directed}
\alias{sim.focal.directed}
\title{Focal sampling directed data}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
sim.focal.directed
}
\value{
a data frame
}
\description{
This is a simulated data frame of directed behaviour collected through focal sampling.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
