% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.focal.undirected.R
\docType{data}
\name{sim.focal.undirected}
\alias{sim.focal.undirected}
\title{Focal sampling undirected data}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
sim.focal.undirected
}
\description{
This is a simulated data frame of undirected behaviour collected through focal sampling.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
