% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.gbi.att.R
\docType{data}
\name{sim.gbi.att}
\alias{sim.gbi.att}
\title{Dyadic attributes for sim.gbi data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 10 columns.
}
\usage{
sim.gbi.att
}
\description{
This is a simulated matrix of dyadic attributes to use as argument M2 in assoc.gfi function.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
