% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.gbi.R
\docType{data}
\name{sim.gbi}
\alias{sim.gbi}
\title{Group by individual (GBI) matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 10 columns.
}
\usage{
sim.gbi
}
\description{
This is a simulated Group by individual (GBI) matrix collected through gambit of the group sampling.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
