% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.grp.R
\docType{data}
\name{sim.grp}
\alias{sim.grp}
\title{Simulated data collected through gambit of the group sampling}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\usage{
sim.grp
}
\value{
a data frame
}
\description{
This is a simulated data set  collected through gambit of the group sampling.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
