% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.m.R
\docType{data}
\name{sim.m}
\alias{sim.m}
\title{Simulated Matrix of interactions}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 20 columns.
}
\usage{
sim.m
}
\value{
a matrix
}
\description{
This is a simulated Matrix of interaction.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
