% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.mcmc.dynamP_function.R
\name{sim.mcmc.dynamP}
\alias{sim.mcmc.dynamP}
\title{This function allows you to simulate data from the following moddel: 
X_t ~ Markov Chain ( X_t-1 , P_t ) [two states]
Y_t ~ Pois (lambda_{X_t})}
\usage{
sim.mcmc.dynamP(tmax, delta.t, start.state = 1, P11, P12, P21, P22, lambda)
}
\arguments{
\item{tmax, }{delta.t, start.state, P11, P12, P21, P22, lambda}
}
\value{
#Output: (1) - x: (1, 2) unobserved two state process 
       (2) - y: (0, ...) "observed" number of interactions at time t
       (3) - N: (0, ...) cumulative count of interactions
       (4) - delta.t
       (5) - t: (0, ...) time (in seconds)
       (6) - 1x1 visual of cumlative counts over time, colored by state
}
\description{
P_1,2_t = (exp(beta.0 + beta.1 * covariate[t])) / 
           (1 + exp(beta.0 + beta.1 * covariate[i])) 
P_1,1_t = 1 - P_1,2_t
P_2,1_t = (exp(alpha)) / (1 + exp(alpha))
P_2,2_t = 1 - P_2,1_t
}
\examples{


lambda = k = c(1, 4)
delta.t = 1 #needs to be 1, else observations dependent on time
sim = sim.mcmc.dynamP(7200, delta.t, start.state = 1, P11, P12, P21, P22, lambda)

}
\keyword{HMM,}
\keyword{Poisson}
\keyword{Process,}
\keyword{covariate,}
\keyword{simulation,}

