% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.mmpp_function.R
\name{sim.mmpp}
\alias{sim.mmpp}
\title{This function allows you to simulate data from the following moddel: 
X_t ~ Markov Chain ( X_t-1 , P ) [two states]
Y_t ~ Pois (lambda_{X_t})}
\usage{
sim.mmpp(tmax, delta.t, start.state = 1, P, lambda)
}
\arguments{
\item{tmax, }{delta.t, start.state, P, lambda}
}
\value{
#Output: (1) - x: (1, 2) unobserved two state process 
       (2) - y: (0, ...) "observed" number of interactions at time t
       (3) - N: (0, ...) cumulative count of interactions
       (4) - delta.t
       (5) - t: (0, ...) time (in seconds)
       (6) - 1x1 visual of cumlative counts over time, colored by state
}
\description{
This function allows you to simulate data from the following moddel: 
X_t ~ Markov Chain ( X_t-1 , P ) [two states]
Y_t ~ Pois (lambda_{X_t})
}
\examples{
P = matrix(c(.99, .01, .01, .99), nrow = 2, byrow = T)
lambda = k = c(1, 4)
delta.t = 1 #needs to be 1, else observations dependent on time
sim = sim.mmpp(7200, delta.t, start.state = 1, P, lambda)

}
\keyword{HMM,}
\keyword{Poisson}
\keyword{Process}
\keyword{simulation,}

