% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.socprog.R
\docType{data}
\name{sim.socprog}
\alias{sim.socprog}
\title{Simulated Socprog data format}
\format{An object of class \code{data.frame} with 100 rows and 3 columns.}
\usage{
sim.socprog
}
\description{
This is a simulated Socprog data format.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{datasets}
