% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_DT_troph}
\alias{sim_DT_troph}
\title{Simulation of Ant Trophallaxis Start Times - Simple Model}
\usage{
sim_DT_troph(time_max, delta_t, start_state = 1, switch_rate = c(0, 0),
  state_tpm, int_rate, num_locations = 1)
}
\arguments{
\item{time_max}{Number of seconds for which to simulate data.}

\item{delta_t}{Length of interval to bin interaction and state 
data into.}

\item{start_state}{Starting trophallaxis rate state (X_1 in model
write - up). Defaults to state 1 (low).}

\item{switch_rate}{Rate at which the rates at which ants trophallax 
switch (between high and low). Denoted by gamma in model write-up. 
Used to calculate probability transition matrix (P).}

\item{state_tpm}{Probably transition matrix for switching between 
trophallaxis states (denoted by P in write-up).}

\item{int_rate}{Rate of starting interactions. Denoted by lambda in 
model write-up).}

\item{num_locations}{Number of chambers/locations in which ants are 
found trophallaxis.}
}
\value{
This function will return the following:
  \enumerate{
    \item inter_persec: (0, ...) 'observed' number of
      interactions per 1 second
    \item state: (1 = low, 2 = high) unobserved two state process
    \item cumu_inter: (0, ...) cumulative count of interactions
    \item bin_inter: Trophallaxis interactions binned into smaller 
      intervals determined by sum over delta_t
    \item bin_state: Trophallaxis rate state binned into smaller 
      intervals, determined by average over delta_t
    \item bin_sec: (0, ...) time (in seconds) binned by delta_t
    \item start_time: vector of interaction start times
    \item location: location of each ant interaction
    \item Visuals: 1x1 visual of cumlative counts over time, colored 
      by state, separated by location if applicable.
  }
}
\description{
\code{sim_DT_troph} simulates number of ant interactions
starting per second and per delta_t seconds.
}
\details{
This function allows you to simulate data from the following
model: X_t ~ Markov Chain ( X_t-1 , P ) [two states] N_t ~
Pois (lambda_L + lambda_Change) lambda_L ~ Gam(a, b)
lambda_Change ~ Gam(c, d) lambda_H = lambda_l + lambda_Change
}
\examples{
P <- matrix(c(.99, .01, .01, .99), nrow = 2, byrow = T)
lambda <- k <- c(1, 4)
delta_t <- 5 
sim <- sim.DT.troph(time_max = 3600, delta_t = delta_t, 
                    start_state = 1, switch_rate = 0, state_tpm = P, 
                    int_rate = lambda, num_locations = 1)

}
\keyword{HMM,}
\keyword{Poisson}
\keyword{Process}
\keyword{simulation,}
