% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_pencov_troph}
\alias{sim_pencov_troph}
\title{Simulation of Ant Trophallaxis Start Times - PenCov Model}
\usage{
sim_pencov_troph(states, time_max, delta_t, start_state = 1, int_rate,
  num_locations = 1, covariate, switch_betas)
}
\arguments{
\item{switch_betas}{}
}
\value{
This function will return the following:
  \enumerate{
    \item inter_persec: (0, ...) 'observed' number of
      interactions per 1 second
    \item state: (1 = low, 2 = high) unobserved two state process
    \item cumu_inter: (0, ...) cumulative count of interactions
    \item bin_inter: Trophallaxis interactions binned into smaller
      intervals determined by sum over delta_t
    \item bin_state: Trophallaxis rate state binned into smaller
      intervals, determined by average over delta_t
    \item bin_sec: (0, ...) time (in seconds) binned by delta_t
    \item start_time: vector of interaction start times
    \item location: location of each ant interaction
    \item Visuals: 1x1 visual of cumlative counts over time, colored
      by state, separated by location if applicable.
      }
}
\description{
Simulation of Ant Trophallaxis Start Times - PenCov Model
}
