% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.ci.R
\name{stat.ci}
\alias{stat.ci}
\title{Confidence interval}
\usage{
stat.ci(x, probs = c(0.05, 0.95))
}
\arguments{
\item{x}{a numeric vector.}

\item{quantile}{numeric vector of probabilities with values in [0,1]. (Values up to 2e-14 outside that range are accepted and moved to the nearby endpoint.)}
}
\value{
a numeric vector of length 2 with the lower and upper confidence interval.
}
\description{
Calculate 25, 50, 95 and 99 Confidence interval of a data set
}
\details{
Confidence interval allow to asses a level of certitude that the average of the 'population' is between a specific range.
}
\author{
Sebastian Sosa,Ivan Puga-Gonzalez
}
\keyword{internal}
