% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.model.diag.R
\name{stat.model.diag}
\alias{stat.model.diag}
\title{Diagnostic plot for lm, glm, lmer,glmer models}
\usage{
stat.model.diag(model)
}
\arguments{
\item{model}{a lm, glm, lmer or glmer object}

\item{formula}{the formula use in the model}
}
\description{
plot Fitted values Vs residuals and residuals qqnorm plot
}
\author{
Sebastian Sosa, Ivan Puga Gonzalez
}
\keyword{internal}
