% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.p.R, R/stat.p.R
\name{stat.p}
\alias{stat.p}
\title{P-value}
\usage{
stat.p(metrics, stat.hist = FALSE)

stat.p(metrics, stat.hist = FALSE)
}
\arguments{
\item{metrics}{an integer vector of permuted statistical values of interest}

\item{stat.hist}{if \emph{true}, a stat.hist of the posterior distribution is ploted. The observed metric is also ploted.}

\item{histogram}{if \emph{true}, an histogram of the posterior distribution is ploted. The observed metric is also ploted.}
}
\value{
\itemize{
\item  An integer vector of left and rigth p-values, if \emph{metrics} is an integer vector and if \emph{histogram} is \emph{false}.
\item  An integer vector of left and rigth p-values and an histogram of the posterior distribution, if \emph{metrics} is an integer vector and if \emph{histogram} is \emph{false}.
}

a numeric vector.

\itemize{
\item  An integer vector of left and rigth p-values, if \emph{metrics} is an integer vector and if \emph{stat.hist} is \emph{false}.
\item  An integer vector of left and rigth p-values and an histogram of the posterior distribution, if \emph{metrics} is an integer vector and if \emph{stat.hist} is \emph{false}.
}
}
\description{
Calculation of the p-value based on a posterior distribution through data permutations.

Calculation of the p-value based on a posterior distribution through data permutations.
}
\details{
Due to data dependency, a valuable p-value in social networking has to be computed through a null model approach.

Due to data dependency, a valuable p-value in social networking has to be computed through a null model approach.
}
\references{
Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.

Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (2018). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.

Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{interanl}
\keyword{internal}
