% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.tauKrPermSig.R
\name{stat.tauKrPermSig}
\alias{stat.tauKrPermSig}
\title{Matrix TauKr correlations}
\usage{
stat.tauKrPermSig(X, Y, nperm = nperm, omitDiag = TRUE)
}
\arguments{
\item{X}{a square matrix of individual interactions or associations}

\item{Y}{a square matrix of individual interactions or attributes}

\item{nperm}{number of permutations.}

\item{omitDiag}{boolean to omit or not the diagonal from the matrices}
}
\value{
a list with the 1) TauKr coefficient value, p-value on the right 2) p-value on the left
}
\description{
Matrix TauKr correlation with significant test.
}
\references{
Hemelrijk, C. K. 1990. Models of, and tests for, reciprocity, unidirectional and other social interaction patterns at a group level. Animal Behavior, 39, 1013-1029
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez.
}
\keyword{internal}
