% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumvis.troph.data_function.R
\name{sumvis.troph.data}
\alias{sumvis.troph.data}
\title{Summary and Visualizations for Ant Trophallaxis Data}
\usage{
sumvis.troph.data(data, entrance, hours, density = "high")
}
\arguments{
\item{data, }{entrance, hours, density}
}
\value{
(1) - forager - number of forager ants
         (2) - forager.ID - ids and number of entrances of each forager
         (3) - allants - total number of ants engaged in trophallaxis
         (4) - all.ID - ids and number of interactions for each ant
         (5a) - Plot - Interactions per Ant with red marked mean
         (5b) - Plot - If low density, interactios per ant by location
         (6) - Plot -  Number of trophallaxis events over time with entrance
           times marked in red.
         (6b) - Plot - If low density, trophallaxis events over time by location
}
\description{
The purpose of this function is to easily summarize and visualize 
ant trophallaxis data along with entrance times of foragers into 
the nest.
}
\examples{
sum = sumvis.troph.data(data = troph.high.4, 
                 entrance = inout.high.4, 
                 hours = 4, 
                 density = "high")
}
