% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_visual_functions.R
\name{sumvis_troph}
\alias{sumvis_troph}
\title{Summary and Visualizations for Ant Trophallaxis Data}
\usage{
sumvis_troph(data, entrance = FALSE, hours, density = "high")
}
\arguments{
\item{data}{Ant trophallaxis data file already loaded and 
prepared with appropriate function.}

\item{entrance}{Data file of (presumably) forager ants 
entering the colony. Used as a covariate in more complex 
models.}

\item{hours}{Hours of observations reflected in trophallaxis 
and entrance data files.}

\item{density}{Options are: \{high, low\} density. Summaries
and plots will also devide data by chamber location in the
"low" density case.}
}
\value{
(1) - forager - number of forager ants (2) - 
  forager.ID - ids and number of entrances of each forager (3)
  - allants - total number of ants engaged in trophallaxis (4)
  - all.ID - ids and number of interactions for each ant (5a) 
  - Plot - Interactions per Ant with red marked mean (5b) - 
  Plot - If low density, interactios per ant by location (6) -
  Plot -  Number of trophallaxis events over time with 
  entrance times marked in red. (6b) - Plot - If low density, 
  trophallaxis events over time by location
}
\description{
The purpose of this function is to easily summarize and 
visualize ant trophallaxis data. In addition, if available, 
entrance times of foragers into the nest will also be 
reflected in the plots.
}
\examples{
sum = sumvis.troph.data(data = troph.high.4, 
                 entrance = inout.high.4, 
                 hours = 4, 
                 density = "high")
}
