% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauSD.R
\name{tauSD}
\alias{tauSD}
\title{TauKr standard deviation}
\usage{
tauSD(X, Y, j, omitDiag)
}
\description{
Matrix TauKr correlationsstandard deviation.
}
\author{
Ivan Puga-Gonzalez, Sebastian Sosa.
}
\keyword{internal}
