% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.ants.fancy_function.R
\name{theme.ants.fancy}
\alias{theme.ants.fancy}
\title{Ants theme for ggplo2}
\usage{
theme.ants.fancy(base_size = 2.5)
}
\value{
A theme element
}
\description{
Import Ants ggplot2 theme. It builds on theme_classic. You can change font family or size.
}
\examples{
qplot(1) + theme.ants.fancy()
}
