% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme.research_function.R
\name{theme.research}
\alias{theme.research}
\title{Research themes for ggplot2}
\usage{
theme.research(theme = "ants", tech_key = list(ants = list(family_title =
  "Circular Air Bold", family_text = "Circular Air Medium", colour_title =
  "#674528", colour_text = "#535353")))
}
\description{
research theme
}
\examples{
theme_tech(theme='ants')
}
\author{
Meridith Bartley
}

