% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.heterogeneity.R
\name{time.heterogeneity}
\alias{time.heterogeneity}
\title{Create a matrix of time of observation per dyades}
\usage{
\method{time}{heterogeneity}(vec)
}
\arguments{
\item{vec}{a numeric vector}
}
\description{
Creat a a matrix of time of observation per dyades in order to control for time of observation
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
