% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tobs_to_mat}
\alias{tobs_to_mat}
\title{Control heterogeneity of time of observation}
\usage{
tobs_to_mat(vec)
}
\description{
repeat a numeric vector n times, where n is the length of this vecctor
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
