% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_char_as_factor}
\alias{vec_char_as_factor}
\title{As factor}
\usage{
vec_char_as_factor(vec)
}
\arguments{
\item{vec}{a vector.}
}
\value{
A vector of same length and order of the input vector.
}
\description{
Transform a vector into a factor vetor.
}
\details{
Permuting a vector can serve to realize node label permutation.
}
\references{
Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods in Ecology and Evolution.

Sosa, S. (\emph{in press}). Social Network Analysis, \emph{in}: Encyclopedia of Animal Cognition and Behavior. Springer.
}
\author{
Sebastian Sosa,Ivan Puga-Gonzales.
}
\keyword{internal}
