% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_char_extract_IdValue}
\alias{vec_char_extract_IdValue}
\title{Extract vector elements}
\usage{
vec_char_extract_IdValue(y, x)
}
\description{
xtract vector elements according to indices
}
\keyword{internal}
