% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_lowertri_to_mat}
\alias{vec_lowertri_to_mat}
\title{Vector to matrix lower triangle}
\usage{
vec_lowertri_to_mat(vec, col, diag)
}
\arguments{
\item{vec}{a numeric vector}

\item{col}{number of columns of the matrix to create}

\item{diag}{doe diagonal is included}
}
\value{
A vector
}
\description{
Create a matrix from a vector
}
\section{Warning}{
 Argument vec must be of the exact same length of the number of cells present in the lower triangle of the amtrix to create.
}

\author{
Sebastian Sosa.
}
\keyword{internal}
