% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_num_extract_IdValue}
\alias{vec_num_extract_IdValue}
\title{Extract ID}
\usage{
vec_num_extract_IdValue(y, x)
}
\description{
Extract ID of element selected by a numeric vector
}
\keyword{internal}
