% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_to_mat}
\alias{vec_to_mat}
\title{Vector to matrix}
\usage{
vec_to_mat(vec, ncol, diag)
}
\arguments{
\item{vec}{a numeric vector}

\item{diag}{a boolean indicating with the argument vec have the diagonal information in it.}

\item{n}{number of columns of the matrix to create}
}
\value{
A square matrix
}
\description{
Create a matrix from a vector
}
\author{
Sebastian Sosa.
}
\keyword{internal}
