% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_to_mat_add_diag}
\alias{vec_to_mat_add_diag}
\title{Vector to matrix and adding diag}
\usage{
vec_to_mat_add_diag(vec, ncol)
}
\arguments{
\item{vec}{a numeric vector}

\item{n}{number of columns of the matrix to create}
}
\value{
A square matrix
}
\description{
Create a matrix from a vector
}
\author{
Sebastian Sosa.
}
\keyword{internal}
