% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.protocol.R
\name{which.protocol}
\alias{which.protocol}
\title{Which metric to choose}
\usage{
which.protocol()
}
\description{
An interactive decision tree to choose the most appropriate ANTs analytical protocol and related function according to the research question.
}
\author{
Sebastian Sosa
}
\keyword{export}
