% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_model_functions.R
\name{DT_pen_mcmc}
\alias{DT_pen_mcmc}
\title{Penalized Discrete Time MCMC Estimation for Trophallaxis data}
\usage{
DT_pen_mcmc(penalty, starts_data, states, ant_file, chamber, hours, a, b, c, d,
  e, f, tau, tau.pen, n_mcmc, delta_t, start, data_out, osa_out, modelsrun_out,
  fig_path, fig_name)
}
\arguments{
\item{penalty}{Penalty parameter for the switching rates
(gamma in model write- up).}

\item{starts_data}{Data file of number of trophallaxis
interactions starting per delta_t time interval.}

\item{states}{Number of unobserved states, generally 2.
("low", "high")}

\item{ant_file}{Data file of ant trophallaxis information,
including Location and start times.}

\item{hours}{Hours of ant observations}

\item{a}{Hyperprior for rates of starting interactions during
"low" state. (lambda_L)}

\item{b}{Hyperprior for rates of starting interactions during
"low" state. (lambda_L)}

\item{c}{Hyperprior for rates of starting interactions during
"high" state. (lambda_H)}

\item{d}{Hyperprior for rates of starting interactions during
"high" state. (lambda_H)}

\item{e}{Hyperprior for rates of starting interactions during "medium" state.}

\item{f}{Hyperprior for the rates of starting interactions during "medium" state.}

\item{tau}{Tuning parameter for the proposal function.}

\item{n_mcmc}{Number of MCMC iteractions.}

\item{delta_t}{Time segemnts the start data is binned into.}

\item{start}{Starting values for the chains.}

\item{data_out}{Name out file name for saving mcmc chains}

\item{osa_out}{Name of out file for saving osa values}

\item{title}{Title of diagnositic plots used to check model
results}
}
\value{
(1) - estimates of X, lambda, gamma, P (2) - 2x2
  visual of estimates over time (runs) (3) - color block state
  switching graph
}
\description{
The purpose of this function is to find MCMC generated
estimates of (1) - the state (X_t = high/low troph rates) of
the colony at time t (2) - the specific rates of interaction
(lambda) of each state (3) - the specific rates of state
switching (gamma) for the ants (3a) - the probability of
moving from one state to another (P matrix) generated from
gamma values.
}
\examples{
theta = matrix(data = c(90, 10, 10, 90), nrow = 2, ncol = 2, byrow = T)
out.high = DT.pen_mcmc.troph()


}
