% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp.to.edgl.R
\name{grp.to.edgl}
\alias{grp.to.edgl}
\title{Linear group by individual data frame to dyadic data frame (edge list)}
\usage{
grp.to.edgl(df, scan, id)
}
\arguments{
\item{df}{a data frame of individual associations.}

\item{scan}{a numeric or character vector representing one or more columns used as scan factors.}

\item{id}{a numeric or character vector indicating the column holding ids of individuals.}
}
\value{
a data frame with individual associations (symmetric data) according to the scan variable(s) declared by the user.
}
\description{
Converts linear group by individual data frame to a dyadic data frame (edge list)
}
\examples{
grp.to.edgl(sim.grp, scan = c('day', 'location'), id = 'ID')
}
\keyword{internal}
