% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnoses_functions.R
\name{penalty_diagnosis}
\alias{penalty_diagnosis}
\title{Diagnoses of Penalized HMM Model}
\usage{
penalty_diagnosis(mcmc_matrix, states, ant_file, chamber, Time, fig_path,
  fig_name, penalty, covariate = NULL)
}
\arguments{
\item{mcmc_matrix}{}

\item{states}{}

\item{ant_file}{}

\item{chamber}{}

\item{Time}{}

\item{fig_path}{}

\item{fig_name}{}

\item{penalty}{}

\item{covariate}{}
}
\value{
Diagnoses plots of parameters in model + state estimates plot. Works with all models.
}
\description{
Diagnoses of Penalized HMM Model
}

