% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.net.weigths.R
\name{perm.net.weigths}
\alias{perm.net.weigths}
\title{Network weigths permutation}
\usage{
perm.net.weigths(M, sym = FALSE, nperm = NULL, progress = TRUE)
}
\arguments{
\item{M}{a square matrix with colmun and row names}

\item{sym}{if \emph{TRUE}, it extracts the lower triangle of the matrix only.}

\item{nperm}{an integer indicating the number of permutations wanted.}

\item{progress}{a boolean indicating the visualization of the permutation process.}
}
\value{
list of square matrices of length nperm + 1.
}
\description{
Permute weigths while keep link structure
}
\examples{
perm.net.weigths(M = sim.m, sym = FALSE, nperm = 10, progress = TRUE)
}
\keyword{internal}
