% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.troph.pairs_function.R
\name{prep.troph.pairs}
\alias{prep.troph.pairs}
\title{Alternative Data Preparation for Ant Trophallaxis Data File}
\usage{
prep.troph.pairs(data)
}
\value{
(1) .
}
\description{
The purpose of this function is to take in .csv files 
and apply necissary changes to the data format. Changes include
}
\examples{
 prep.troph.pairs(high4)


}

