% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo.perm.ds.grp.R
\name{redo.ds.grp}
\alias{redo.ds.grp}
\title{gambit of the group data stream function for error found in permutations}
\usage{
redo.ds.grp(
  family,
  new.perm,
  gbi,
  oda,
  odf,
  target.metrics,
  formula,
  Scan,
  method,
  ctrlf,
  fam,
  ...
)
}
\description{
Performs cumulative data stream permutations for gambit of the group until the glmm found no error or warnings
}
\keyword{internal}
