% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_visual_functions.R
\name{sumtable_model}
\alias{sumtable_model}
\title{Post MCMC Summary Table and Visuals}
\usage{
sumtable_model(results, compare, file_path, file_name, model)
}
\arguments{
\item{file_name}{}
}
\description{
Post MCMC Summary Table and Visuals
}
