% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.heterogeneity.R
\name{time.heterogeneity}
\alias{time.heterogeneity}
\title{Control heterogeneity of time of observation}
\usage{
\method{time}{heterogeneity}(vec)
}
\description{
repeat a numeric vector n times, where n is the length of this vecctor
}
\author{
Sebastian Sosa, Ivan Puga-Gonzalez
}
\keyword{internal}
