% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.metric.R
\name{which.metric}
\alias{which.metric}
\title{Which metric to choose}
\usage{
which.metric()
}
\value{
a decision tree.
}
\description{
An interactive decision tree to choose the most appropriate network measure according to the research question and network analysis level.
}
\details{
For more details on each of these metrics, see Sosa et al. 2020.
}
\references{
Sosa, S., Sueur, C., & Puga Gonzalez, I. (2020). Network measures in animal social network analysis: their strenths, limits, interpretations and uses. Methods in Ecology and Evolution.
}
\author{
Sebastian Sosa
}
\keyword{export}
